fix typo

+++ models/model.py
@@ -107,7 +107,7 @@ class AffWild2VA(pl.LightningModule):
     
     def ce_loss(self, y_hat, y, mask):
         loss = F.cross_entropy(y_hat.view(-1, y_hat.size(-1)), y.view(-1), reduction='none')
-        return (loss * mask.view(-1)).mean()
+        return (loss * mask.view(-1).float()).mean()
     
     def mse_loss(self, y_hat, y):
         return F.mse_loss(y_hat, y)
